% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaps-data.R
\name{add_utm_tracks_yaps}
\alias{add_utm_tracks_yaps}
\title{convert a dataframe of gps track points into UTM}
\usage{
add_utm_tracks_yaps(
  track,
  starttime,
  endtime,
  utm_zone,
  crs = 4326,
  formatts = TRUE
)
}
\arguments{
\item{track}{one dataframe file with all track points from a SINGLE track. Best practice recommendation is to import GPX data using plotKML's readGPX function, and extract the desired trackfrom the resulting list}

\item{utm_zone}{UTM zone in which the field site is located. "18N" for Lac-Saint-Pierre, and "17N" for James Bay sites}

\item{crs}{4-digit EPSG code for the coordinate system the tracks were recorded in. Default is 4326 for EPSG:WGS84.}

\item{formatts}{boolean term for whether the}

\item{starttime/endtime}{character string or POSIXt date and time stamp to filter only a given date range of data}
}
\value{
returns a dataframe with the UTM coordinates added
}
\description{
convert a dataframe of gps track points into UTM
}
