% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaps-data.R
\name{add_utm_wpts_yaps}
\alias{add_utm_wpts_yaps}
\title{convert a dataframe of gps waypoints into UTM}
\usage{
add_utm_wpts_yaps(gps_df, utm_zone, crs = 4326)
}
\arguments{
\item{gps_df}{one dataframe file with all gps waypoints to format. Best practice recommendation is to import GPX data using plotKML's readGPX function, and extract the waypoints dataframe fromt he resulting list}

\item{utm_zone}{UTM zone in which the field site is located. "18N" for Lac-Saint-Pierre, and "17N" for James Bay sites}

\item{crs}{4-digit EPSG code for the coordinate system gps points are currently in. Default is 4326 for EPSG:WGS84.}
}
\value{
returns a dataframe with the UTM coordinates added
}
\description{
convert a dataframe of gps waypoints into UTM
}
