% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capture_outcomes}
\alias{capture_outcomes}
\title{Approved fish capture outcome codes}
\format{
A vector of the entries "released", "tagged",
"euthenizedCapture", "euthenizedSurgery","euthenizedRecovery",
"mortalityCapture" "mortalitySurgery", "mortalityRecovery"
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}}
}
\usage{
capture_outcomes
}
\description{
Approved fish capture outcome codes
}
\keyword{datasets}
