% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_angling}
\alias{check_angling}
\title{Performs comprehensive verification on angling database}
\usage{
check_angling(angling)
}
\arguments{
\item{angling}{the equipment log tibble}
}
\value{
an angling tibble with the added data flag column
}
\description{
Performs comprehensive verification on angling database
}
