% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_cast_netting}
\alias{check_cast_netting}
\title{Performs comprehensive verification on cast_netting database}
\usage{
check_cast_netting(cast_netting)
}
\arguments{
\item{cast_netting}{the equipment log tibble}
}
\value{
an cast_netting tibble with the added data flag column
}
\description{
Performs comprehensive verification on cast_netting database
}
