% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_equipment_log}
\alias{check_equipment_log}
\title{Performs comprehensive verification on equipment_log}
\usage{
check_equipment_log(equipment_log)
}
\arguments{
\item{equipment_log}{the equipment log tibble}
}
\value{
an equipment_log tibble with the added data flag column
}
\description{
Performs comprehensive verification on equipment_log
}
