% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_fish}
\alias{check_fish}
\title{Performs comprehensive verification on fish database}
\usage{
check_fish(fish)
}
\arguments{
\item{fish}{the equipment log tibble}
}
\value{
an fish tibble with the added data flag column
}
\description{
Performs comprehensive verification on fish database
}
