% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_folderlocation}
\alias{check_folderlocation}
\title{Checks that data quality control is done in the right place}
\usage{
check_folderlocation(testing_mode = FALSE)
}
\description{
Checks that data quality control is done in the right place
}
