% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_fyke}
\alias{check_fyke}
\title{Performs comprehensive verification on fyke database}
\usage{
check_fyke(fyke)
}
\arguments{
\item{fyke}{the equipment log tibble}
}
\value{
an angling tibble with the added data flag column
}
\description{
Performs comprehensive verification on fyke database
}
