% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_gps_records}
\alias{check_gps_records}
\title{Performs comprehensive verification on gps_records database}
\usage{
check_gps_records(gps_records)
}
\arguments{
\item{gps_records}{the equipment log tibble}
}
\value{
an gps_records tibble with the added data flag column
}
\description{
Performs comprehensive verification on gps_records database
}
