% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_range_test}
\alias{check_range_test}
\title{Performs comprehensive verification on range_test database}
\usage{
check_range_test(range_test)
}
\arguments{
\item{range_test}{the equipment log tibble}
}
\value{
an range_test tibble with the added data flag column
}
\description{
Performs comprehensive verification on range_test database
}
