% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{condition_codes}
\alias{condition_codes}
\title{Approved fish condition codes}
\format{
a vector with entries "vigorous", "normal", "weak"
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}}
}
\usage{
condition_codes
}
\description{
Approved fish condition codes
}
\keyword{datasets}
