% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_capture_method}
\alias{.check_capture_method}
\title{Checks to make sure that the capture method one of the accepted values and is filled in}
\usage{
.check_capture_method(capture_method)
}
\arguments{
\item{capture_method}{a single station ID entry entry of the equipment type (as a character)}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure that the capture method one of the accepted values and is filled in
}
