% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_capture_outcome}
\alias{.check_capture_outcome}
\title{Checks outcome of fish capture}
\usage{
.check_capture_outcome(outcome, mandatory = TRUE)
}
\arguments{
\item{outcome}{a single equipment entry (as a character string)}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for each entry
}
\description{
Checks outcome of fish capture
}
