% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_condition}
\alias{.check_condition}
\title{Checks whether the conditions code has been entered}
\usage{
.check_condition(condition, mandatory = TRUE)
}
\arguments{
\item{condition}{a single condition value to be checked}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Checks whether the conditions code has been entered
}
