% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_crew}
\alias{.check_crew}
\title{Checks to make sure that the crew is entered and comma separated}
\usage{
.check_crew(crew, mandatory = TRUE)
}
\arguments{
\item{crew}{a single station ID entry entry of the equipment type (as a character)}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure that the crew is entered and comma separated
}
