% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_deploy}
\alias{.check_deploy}
\title{Checks to make sure the deployment type is entered and is one of the permitted entries}
\usage{
.check_deploy(deploy, site)
}
\arguments{
\item{deploy}{a single station ID entry entry of the equipment type (as a character)}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the deployment type is entered and is one of the permitted entries
}
