% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_distance}
\alias{.check_distance}
\title{Checks to make sure the depth is within sensible bounds}
\usage{
.check_distance(dist, mandatory = FALSE)
}
\arguments{
\item{dist}{a single numeric distance value}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the depth is within sensible bounds
}
