% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_dna_scale_id}
\alias{.check_dna_scale_id}
\title{Checks to make sure the weight is within sensible bounds}
\usage{
.check_dna_scale_id(dna_id, scale_id, mandatory = FALSE)
}
\arguments{
\item{scale_id}{a single scale sample ID number}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the weight is within sensible bounds
}
