% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{.check_equip_serial_match}
\alias{.check_equip_serial_match}
\title{Checks to make sure the serial ID number matches the equipment type}
\usage{
.check_equip_serial_match(equip, serial)
}
\arguments{
\item{equip}{a single equipment type entry of the equipment type (as a character)}

\item{serial}{a single serial ID entry of the serial id number (as a character or integer)}
}
\value{
a report of the check status for each entry
}
\description{
Checks to make sure the serial ID number matches the equipment type
}
