% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_fykeid}
\alias{.check_fykeid}
\title{Checks to make sure that the capture method one of the accepted values and is filled in}
\usage{
.check_fykeid(fykeid, mandatory = FALSE)
}
\arguments{
\item{fykeid}{a single fyke id number entry}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure that the capture method one of the accepted values and is filled in
}
