% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{.check_gps_site_match}
\alias{.check_gps_site_match}
\title{Checks to make sure the gps latitude and longitude match the site ID
NOT FUNCTIONALL****}
\usage{
.check_gps_site_match(site, lat, lon)
}
\arguments{
\item{site}{site ID}

\item{lat}{a single numeric latitude value}

\item{lon}{a single numeric longitude value}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the gps latitude and longitude match the site ID
NOT FUNCTIONALL****
}
