% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_lat}
\alias{.check_lat}
\title{Checks to make sure the gps latitude is within sensible bounds}
\usage{
.check_lat(lat)
}
\arguments{
\item{lat}{a single numeric latitude value}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the gps latitude is within sensible bounds
}
