% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_lon}
\alias{.check_lon}
\title{Checks to make sure the gps longitude is within sensible bounds}
\usage{
.check_lon(lon)
}
\arguments{
\item{lon}{a single numeric longitude value}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the gps longitude is within sensible bounds
}
