% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_net_action}
\alias{.check_net_action}
\title{Checks to make sure the sex is entered as on of the approvd codes}
\usage{
.check_net_action(net_action, mandatory = TRUE)
}
\arguments{
\item{net_action}{a single character string sex and maturity code}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the sex is entered as on of the approvd codes
}
