% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_nrods_or_nets}
\alias{.check_nrods_or_nets}
\title{Checks to make sure the length is within sensible bounds}
\usage{
.check_nrods_or_nets(nrods, mandatory = TRUE, fishtype = c("nrods", "nnets"))
}
\arguments{
\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}

\item{length}{a single numeric length value}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the length is within sensible bounds
}
