% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_rt}
\alias{.check_rt}
\title{Checks to make sure the deployment type is entered and is one of the permitted entries}
\usage{
.check_rt(rt, mandatory = TRUE)
}
\arguments{
\item{rt}{a single rt type code}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the deployment type is entered and is one of the permitted entries
}
