% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_sex}
\alias{.check_sex}
\title{Checks to make sure the sex is entered as on of the approvd codes}
\usage{
.check_sex(sex)
}
\arguments{
\item{sex}{a single character string sex and maturity code}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the sex is entered as on of the approvd codes
}
