% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_single_initials}
\alias{.check_single_initials}
\title{Checks to make sure that entries that require single initials are correctly filled out}
\usage{
.check_single_initials(initials)
}
\arguments{
\item{initials}{a single 2 or 3-letter initials code}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure that entries that require single initials are correctly filled out
}
