% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_stnid}
\alias{.check_stnid}
\title{Checks to make sure the station ID follows a valid format}
\usage{
.check_stnid(stnid, deploy = NA)
}
\arguments{
\item{stnid}{a single station ID entry entry of the equipment type (as a character)}

\item{deploy}{a single station ID entry entry of the equipment type (as a character)}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the station ID follows a valid format
}
