% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_temp}
\alias{.check_temp}
\title{Checks to make sure the temp is within sensible bounds}
\usage{
.check_temp(temp)
}
\arguments{
\item{temp}{a single numeric temp value}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the temp is within sensible bounds
}
