% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_time}
\alias{.check_time}
\title{Checks manually-input time entries}
\usage{
.check_time(time)
}
\arguments{
\item{time}{a single time entry (as a character string)}
}
\value{
a report of the check status
}
\description{
Checks manually-input time entries
}
