% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_weight}
\alias{.check_weight}
\title{Checks to make sure the weight is within sensible bounds}
\usage{
.check_weight(weight)
}
\arguments{
\item{weight}{a single numeric weight value}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the weight is within sensible bounds
}
