% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_wpt}
\alias{.check_wpt}
\title{Checks to make sure the waypoints are entered in an acceptable format}
\usage{
.check_wpt(
  wpt,
  wpt_type = c("equipment", "capture", "release"),
  mandatory = FALSE
)
}
\arguments{
\item{wpt}{a single entry of waypoint names as a comma-separated string}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Currently checks the following info:
\enumerate{
\item if waypoints are given by a comma-separated list, are there any missing values?
\item do waypoints start with a lower-case 'd' or 'w' (corresponding to Darwin or Wallace units)
\item waypoints should be 3-4 characters long
}
}
