% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_current_template}
\alias{download_current_template}
\title{Downloads the most recent template file to use for data entry}
\usage{
download_current_template(folder_path)
}
\arguments{
\item{folder_path}{the path of the folder to download the template to.}
}
\description{
Downloads the most recent template file to use for data entry
}
