% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equip_types}
\alias{equip_types}
\title{Approved equipment type codes}
\format{
A vector of the entries "VR2Tx", "VR2W",  "tag", "RTtag", "VR100", "AMchl", "AMdo",  "AMsal", "GPS", "other"
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}}
}
\usage{
equip_types
}
\description{
Approved equipment type codes
}
\keyword{datasets}
