% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equipment_actions}
\alias{equipment_actions}
\title{Approved equipment action codes}
\format{
A vector of the entries "On", "Off", "deployed", "retrieved", "moved","notLocated", "notRetrieved", "located",
"dataDownload", "checked", "configure", "recharge",  "anchorMid", "anchorEnd",
"repairSent", "repairReturn", and
"other"
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}}
}
\usage{
equipment_actions
}
\description{
Approved equipment action codes
}
\keyword{datasets}
