% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-extraction.R
\name{find_equipment_event}
\alias{find_equipment_event}
\title{Find the final recorded record of a given set of IDs from a equipment_log object}
\usage{
find_equipment_event(
  log_data,
  ID,
  eventtype = NULL,
  type = c("first", "last", "all")
)
}
\arguments{
\item{log_data}{An object of the "equipment_log" class containing sequential records of equipment}

\item{ID}{a character vector of IDs to load}

\item{eventtype}{dunno}

\item{type}{dunno}
}
\value{
An equipment_log tibble containing the final recorded event for each ID, or NA if the ID did not occur in the data
}
\description{
Find the final recorded record of a given set of IDs from a equipment_log object
}
