% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaps-data.R
\name{format_receiverlog_yaps}
\alias{format_receiverlog_yaps}
\title{Reformats the receiver log into a format that interfaces well with yaps. The receiver log is downloaded from the VUE software database}
\usage{
format_receiverlog_yaps(receiverLog)
}
\arguments{
\item{receiverLog}{the VUE receiver detections database}
}
\value{
returns a reformated detections database.
}
\description{
Reformats the receiver log into a format that interfaces well with yaps. The receiver log is downloaded from the VUE software database
}
