% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fyke_actions}
\alias{fyke_actions}
\title{Approved fyke action codes}
\format{
A vector of the entries "set", "retrieved", "checked", "moved"
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}}
}
\usage{
fyke_actions
}
\description{
Approved fyke action codes
}
\keyword{datasets}
