% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{import_database_xl}
\alias{import_database_xl}
\title{Imports the excel database into R, as a list of multiple tibbles}
\usage{
import_database_xl(path)
}
\arguments{
\item{path}{a string designating the directory path of the database excel sheet}
}
\value{
a report of the check status for the entry
}
\description{
Imports the excel database into R, as a list of multiple tibbles
}
