% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_numeric_like}
\alias{is_numeric_like}
\title{Test whether a vector is numeric or coercible to numeric}
\usage{
is_numeric_like(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\value{
Logical (length 1).
}
\description{
\code{is_numeric_like()} tests whether an object is "coercible to
numeric" by the particular standards of scrutiny. This means:
\itemize{
\item Integer and double vectors are \code{TRUE}.
\item Logical vectors are \code{FALSE}, as are non-vector objects.
\item Other vectors (most likely strings) are \code{TRUE} if all their non-\code{NA}
values can be coerced to non-\code{NA} numeric values, and \code{FALSE} otherwise.
\item Factors are first coerced to string, then tested.
\item Lists are tested like atomic vectors unless any of their elements have
length greater 1, in which case they are always \code{FALSE}.
\item If all values are non-numeric, non-logical \code{NA}, the output is also \code{NA}.
}

This code was adapted from the 'scrutiny' R package
(https://cran.r-project.org/web/packages/scrutiny/index.html) and is used
under the terms of the MIT+ license.
}
\details{
\code{is_numeric_like()} returns \code{FALSE} for logical vectors simply because
these are displayed as strings, not as numbers, and the usual coercion
rules would be misleading in this context. Likewise, the function treats
factors like strings because that is how they are displayed: the fact that
factors are stored as integers is irrelevant.

Why store numbers as strings or factors? Only these data types can preserve
trailing zeros, and only if the data were originally entered as strings.
See \code{vignette("wrangling")}, section \emph{Trailing zeros}.
}
