% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{merge_waypoints}
\alias{merge_waypoints}
\title{Merges GPS waypoints from a GPX file with waypoints input in the Equipment Log}
\usage{
merge_waypoints(folder_path, remerge = F)
}
\arguments{
\item{folder_path}{the path of the raw folder in which the data entry xlsx sheets and .GPX file for merging are stored.}

\item{remerge}{TRUE/FALSE whether the waypoint data is to be re-merged (TRUE) after addressing waypoint flags or merged for the first time (FALSE)}
}
\description{
this function imports a GPX file stored in the "raw" folder and imports latitude and longitude from matching waypoints into the flagged version of the equipment_log, while adding flags to the flagged equipment_log.csv for potential issues with GPS points.
This function should be run after check_dataentry once all non-GPS flags that can be addressed have been addressed.
}
