% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otn-data.R
\name{otn_instrument_deployment}
\alias{otn_instrument_deployment}
\title{Exports a csv with tagging data formatted for the OTN metadata sheet}
\usage{
otn_instrument_deployment(
  equipmentlog,
  sink,
  startdate = "1970-01-01",
  enddate = "3000-01-01",
  sites = c("JEA", "JGR", "JMA", "JRU", "JOY", "LEA", "LWE"),
  equipment_types = c("VR2Tx", "VR2W", "RTtag", "AMchl", "AMdo", "AMsal"),
  lead_tech = NULL
)
}
\arguments{
\item{equipmentlog}{equipmentlog database to extract data from}

\item{sink}{the filepath where to save the csv file}

\item{sites}{(optional) the site codes to filter for}

\item{equipment_types}{(optional) the equipment to filter for}

\item{lead_tech}{Last Name, First initial of lead technician responsible for deployment of these receivers}

\item{startdate/enddate}{Character string or POSIXt date and time stamp to filter only a given date range of data}
}
\value{
returns a reformated detections database.
}
\description{
This is useful for uploading data to Fathom Central and sharing data with OTN
}
