% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otn_instrument_deployment_headers}
\alias{otn_instrument_deployment_headers}
\title{OTN metadata deployment sheets headers}
\format{
a vector with character data for all headers in the OTN deployment metadata sheet
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}. Based off Instrument Deployment Metadata data sheet provided by OTN: https://members.oceantrack.org/data/data-collection.}
}
\usage{
otn_instrument_deployment_headers
}
\description{
OTN metadata deployment sheets headers
}
\keyword{datasets}
