% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otn-data.R
\name{otn_tagging_metadata}
\alias{otn_tagging_metadata}
\title{Exports a csv with tagging data formatted for the OTN metadata sheet}
\usage{
otn_tagging_metadata(
  fish,
  sink,
  startdate = "1970-01-01",
  enddate = "3000-01-01",
  sp = c("COAR", "COCL", "ESLU", "MOHU", "MOSA", "OSMO", "PEFL", "SAFO", "SANA", "SAVI",
    "TITI"),
  tag_models = NULL,
  tagger = NULL
)
}
\arguments{
\item{fish}{fish database to extract data from}

\item{sink}{the filepath where to save the csv file}

\item{sp}{(optional) the species code to filter for}

\item{tag_models}{(optional) the tag model to filter for}

\item{tagger}{(optional) the tagger/surgeon to filter for (as a single initial or as a vector of initials)}

\item{startdate/enddate}{Character string or POSIXt date and time stamp to filter only a given date range of data}
}
\value{
returns a reformated detections database.
}
\description{
This is useful for uploading data to Fathom Central and sharing data with OTN
}
