% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otn_tagging_metadata_headers}
\alias{otn_tagging_metadata_headers}
\title{OTN metadata tagging sheets headers}
\format{
a vector with character data for all headers in the OTN tagging metadata sheet
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}. Based off Tag Metadata data sheet provided by OTN: https://members.oceantrack.org/data/data-collection.}
}
\usage{
otn_tagging_metadata_headers
}
\description{
OTN metadata tagging sheets headers
}
\keyword{datasets}
