% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_control_codes}
\alias{quality_control_codes}
\title{Provides a summary of the possible error codes and their meaning}
\usage{
quality_control_codes()
}
\description{
Provides a summary of the possible error codes and their meaning
}
