% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_serial_id}
\alias{reference_serial_id}
\title{Reference document of Innovasea-Provided equipment}
\format{
A data frame with 82 rows and 3 columns:

\describe{
\item{Type}{Type of equipment (based on our defined codeing system)}
\item{ProductID}{Official Innovasea Product ID}
\item{SerialNo}{serial number of each piece of equipment}
}
}
\source{
{LTMFTN database of Innovasea Equipment}
}
\usage{
reference_serial_id
}
\description{
List of all innovasea-provided equipment, equipment type, product ID and serial number. \strong{\strong{MUST BE UPDATED WHEN NEW EQUIPMENT IS PURCHASED}}
}
\keyword{datasets}
