% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex_options}
\alias{sex_options}
\title{Approved fish sex and maturity codes}
\format{
A vector of the entries "MM", "IM", "MF", "IF", "unk"
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}}
}
\usage{
sex_options
}
\description{
Approved fish sex and maturity codes
}
\keyword{datasets}
