% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sites}
\alias{sites}
\title{Approved work sites code}
\format{
A vector of the entries "JEA", "JWE", "JWA", JWH", "JGR", "JMA", "JRU", "JOY", "LEA", "LWE", "LSW", "base", "lab", and "other"
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}}
}
\usage{
sites
}
\description{
Approved work sites code
}
\keyword{datasets}
