% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{Approved fish species codes '}
\format{
A vector of the entries 'c("COAR",  "COCL", "ESLU",
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}}
}
\usage{
species_codes
}
\description{
Approved fish species codes '
}
\keyword{datasets}
