% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{track_gpx}
\alias{track_gpx}
\title{GPS track data frame}
\format{
A data frame with 8323 rows and 4 columns:

\describe{
\item{lon, lat}{longitude and latitude in WGS84}
\item{ele}{elevation from sea level  in meters}
\item{time}{time of data download in YYYY-MM-DDTHH:MM:SSZ}
}
}
\source{
{From an eTREX device, data imported into GARMIN basecamp, downloaded as a GPX and loaded into R using readGPX()}
}
\usage{
track_gpx
}
\description{
An example of  GPS track data imported from a GPX file using readGPX(). This is used to illustrate GPS data cleaning steps necessary for analysis
}
\keyword{datasets}
