% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{visualize_data_check}
\alias{visualize_data_check}
\title{Appends the checked data entry files to the relevant csv databases}
\usage{
visualize_data_check(flagged_folder_path = NULL, database = NULL)
}
\arguments{
\item{flagged_folder_path}{the filepath to the flagged folder where files to be examined are kept}

\item{database}{the database object to be checked. Alternate to folder_path, to be used only within check_dataentry() function}
}
\value{
visual inspection aids to help identify error that are not captured by quality control
}
\description{
Appends the checked data entry files to the relevant csv databases
}
