% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vue_det}
\alias{vue_det}
\title{Raw VUE detection data}
\format{
A data frame with 1000 rows and 13 columns:

\describe{
\item{Date.UTC., Time.UTC.}{Date and time of dection in UTC}
\item{Receiver}{which receiver recorded the detection}
\item{Transmitter}{which transmitter code was detected}
\item{TransmitterName, TransmitterSerial}{Unpopulated columns (not applicable for our project)}
\item{SensorValue}{Untransformed interger sensor values acoustically transmitted by an AquaMeasure sensor}
\item{SensorUnit}{}
\item{StationName}{Station Name assigned tot he receiver at the time of recording}
\item{Latitude, Longitude}{Unpopulated columns (not applicable for our project)}
\item{TransmitterType, SensorPrecision}{Unpopulated Columns for sensor tags (not applicable for our project)}
}
}
\source{
{subset of VUE download}
}
\usage{
vue_det
}
\description{
An example of a raw VUE download of detection data. This is used to illustrate data cleaning steps necessary for analysis
}
\keyword{datasets}
