% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wpts_gpx}
\alias{wpts_gpx}
\title{GPS waypoint data frame}
\format{
A data frame with 82 rows and 8 columns:

\describe{
\item{lon, lat}{longitude and latitude in WGS84}
\item{ele}{elevation from sea level  in meters}
\item{time}{time of data download in YYYY-MM-DDTHH:MM:SSZ}
\item{name}{the ID name assigned to each waypoint}
\item{sym}{symbol type used by GARMIN}
\item{type, extensions}{GARMIN extra data}
}
}
\source{
{From an eTREX device, data imported into GARMIN basecamp, downloaded as a GPX and loaded into R using readGPX()}
}
\usage{
wpts_gpx
}
\description{
An example of a GPS waypoint data imported from a GPX file using readGPX(). This is used to illustrate GPS data cleaning steps necessary for analysis
}
\keyword{datasets}
