% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{append_to_database}
\alias{append_to_database}
\title{Appends the checked data entry files to the relevant csv databases}
\usage{
append_to_database(folder_path)
}
\arguments{
\item{folder_path}{the new data entry database that has been checked and is ready to add to the main database}
}
\value{
a report of the check status for the entry
}
\description{
Appends the checked data entry files to the relevant csv databases
}
