% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capture_methods}
\alias{capture_methods}
\title{Approved fish capture method codes}
\format{
A vector of the entries "fyke", "angling", "castNet", "other"
}
\source{
{manually created from the excel data validation}
}
\usage{
capture_methods
}
\description{
Approved fish capture method codes
}
\keyword{datasets}
