% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_dataentry}
\alias{check_dataentry}
\title{Performs comprehensive verification of all excel data entry files within a given folder}
\usage{
check_dataentry(
  folder_path,
  return_summary = TRUE,
  recheck = FALSE,
  visualize = TRUE
)
}
\arguments{
\item{folder_path}{the path of the folder in which the data entry xlsx sheets are stored. Folder must only contain the correct format of data entry sheets}

\item{return_summary}{whether to output a summary of data checking to the console}

\item{recheck}{TRUE/FALSE whether the data is to be rechecked (TRUE) or checked for the first time (FALSE)}

\item{visualize}{whether you want to return visual items for further data checking}
}
\value{
if visualize = TRUE, this function returns a list of visuals for the data checker to go through
}
\description{
this function imports excel documents stored in the "raw" folder that need quality control. It performs a comprehensive quality control, and saves csv files output to the "flagged" folder for the QC person to go through and correct the data flags.
}
