% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deploy_types}
\alias{deploy_types}
\title{Approved deployment type codes}
\format{
A vector of the entries "GA", "GR", "RT","SE", and  "other"
}
\source{
{Running the "make-validation-datafiles.R" script in \code{data-raw}}
}
\usage{
deploy_types
}
\description{
Approved deployment type codes
}
\keyword{datasets}
