% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deploy_types}
\alias{deploy_types}
\title{Approved deployment type codes}
\format{
A vector of the entries "GA", "GR", "RT", "other"
}
\source{
{manually created from the excel data validation}
}
\usage{
deploy_types
}
\description{
Approved deployment type codes
}
\keyword{datasets}
