% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{.check_crew}
\alias{.check_crew}
\title{Checks to make sure that the crew is entered and comma separated}
\usage{
.check_crew(crew)
}
\arguments{
\item{crew}{a single station ID entry entry of the equipment type (as a character)}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure that the crew is entered and comma separated
}
