% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_depth}
\alias{.check_depth}
\title{Checks to make sure the depth is within sensible bounds}
\usage{
.check_depth(depth)
}
\arguments{
\item{depth}{a single numeric depth value}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the depth is within sensible bounds
}
