% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{.check_equip}
\alias{.check_equip}
\title{Checks equipment type to make sure it is within the list of allowed values}
\usage{
.check_equip(equip)
}
\arguments{
\item{site}{a single date entry (as a character string)}
}
\value{
a report of the check status for each entry
}
\description{
Checks equipment type to make sure it is within the list of allowed values
}
