% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_length}
\alias{.check_length}
\title{Checks to make sure the length is within sensible bounds}
\usage{
.check_length(length, mandatory = TRUE)
}
\arguments{
\item{length}{a single numeric length value}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure the length is within sensible bounds
}
