% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_serial}
\alias{.check_serial}
\title{Checks the serial ID to make sure it is listed in our equipment list}
\usage{
.check_serial(serial)
}
\arguments{
\item{serial}{a single serial id entry (as a character or integer)}
}
\value{
a report of the check status for each entry
}
\description{
Checks the serial ID to make sure it is listed in our equipment list
}
