% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{.check_site}
\alias{.check_site}
\title{Checks site ID to make sure it is within the list of allowed values}
\usage{
.check_site(site)
}
\arguments{
\item{site}{a single site entry (as a character string)}
}
\value{
a report of the check status for each entry
}
\description{
Checks site ID to make sure it is within the list of allowed values
}
