% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_species}
\alias{.check_species}
\title{Checks to make sure that the fish species is of the accepted values and is filled in}
\usage{
.check_species(species, mandatory = TRUE)
}
\arguments{
\item{species}{a single station ID entry entry of the equipment type (as a character)}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
Checks to make sure that the fish species is of the accepted values and is filled in
}
