% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_internal_fncs.R
\name{.check_wpt}
\alias{.check_wpt}
\title{Checks to make sure the waypoints are entered in an acceptable format}
\usage{
.check_wpt(wpt, mandatory = FALSE)
}
\arguments{
\item{wpt}{a single entry of waypoint names as a comma-separated string}

\item{mandatory}{TRUE/FALSE designates if the data to be checked is mandatory}
}
\value{
a report of the check status for the entry
}
\description{
CURRENTLY NOT CHECKING ANYTHING> POSSIBLE OPTIONS: checking for the presence of special cahracters, checking from proper comma delimitations
}
