% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equipment_actions}
\alias{equipment_actions}
\title{Approved equipment action codes}
\format{
A vector of the entries "On", "Off", "deployed", "retrieved", "moved" , "dataDownload", "checked", "configure", "recharge", "other"
}
\source{
{manually created from the excel data validation}
}
\usage{
equipment_actions
}
\description{
Approved equipment action codes
}
\keyword{datasets}
