% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otn_instrument_deployment_headers}
\alias{otn_instrument_deployment_headers}
\title{OTN metadata deployment sheets headers}
\format{
a vector with character data for all headers in the OTN deployment metadata sheet
}
\source{
{otn_instrument_deployment_headers <- c("OTN_ARRAY",	"STATION_NO",	"DEPLOY_DATE_TIME",	"DEPLOY_LAT",	"DEPLOY_LONG",	"BOTTOM_DEPTH",	"RISER_LENGTH",	"INSTRUMENT_DEPTH",	"INS_MODEL_NO",	"INS_SERIAL_NO"	,"CODE_SET"	,"TRANSMITTER",	"TRANSMIT_MODEL",	"AR_MODEL_NO",	"AR_SERIAL_NO",	"DEPLOYED_BY" ,	"RECOVERED" ,	"RECOVER_DATE_TIME" , "RECOVER_LAT" ,	"RECOVER_LONG" ,"DATA_DOWNLOADED" ,	"DOWNLOAD_DATE_TIME" ,"FILENAME",	"COMMENTS")					usethis::use_data(otn_instrument_deployment_headers)  }
}
\usage{
otn_instrument_deployment_headers
}
\description{
OTN metadata deployment sheets headers
}
\keyword{datasets}
