% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rt_types}
\alias{rt_types}
\title{Approved rt type codes}
\format{
a vector with entries "boat", "fixed_st", "fixed_lt"
}
\source{
{rt_types <- c("boat", "fixed_st", "fixed_lt")           usethis::use_data(rt_types)}
}
\usage{
rt_types
}
\description{
Approved rt type codes
}
\keyword{datasets}
