% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{Approved fish species codes}
\format{
A vector of the entries 'c("COAR",  "COCL", "ESLU",  "MOHU",  "MOSA" , "OSMO", "PEFL", "SAFO", "SANA", "SAVI",  "TITI", "bycatch")
}
\source{
{manually created from the excel data validation}
}
\usage{
species_codes
}
\description{
Approved fish species codes
}
\keyword{datasets}
