% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{Approved fish species codes}
\format{
A vector of the entries "BKTR", "CISC", "CPRD", "LKTR", "LKWH", "NRPK", "RNSM", "STBS", "TENC", "WALL", "YLPR", "bycatch"
}
\source{
{manually created from the excel data validation}
}
\usage{
species_codes
}
\description{
Approved fish species codes
}
\keyword{datasets}
