% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tag_models}
\alias{tag_models}
\title{Approved tag model codes}
\format{
a vector with entries "V6", "V7", "V8", "V9", "V13", "V16"
}
\source{
{tag_models <- c("V6", "V7", "V8", "V9", "V13", "V16")       usethis::use_data(tag_models)}
}
\usage{
tag_models
}
\description{
Approved tag model codes
}
\keyword{datasets}
